// extensions/bangumi/src/types/generated.ts
export interface components {
  schemas: {
// ...
    Character: {
      /** ID */
      id: number
      /** Name */
      name: string
      /** @description 角色，机体，舰船，组织... */
      type: number & components["schemas"]["CharacterType"]
      /**
       * Images
       * @description object with some size of images, this object maybe `null`
       */
      images?: components["schemas"]["PersonImages"]
      /** Summary */
      summary: string
      /** Locked */
      locked: boolean
      /**
       * Infobox
       * @description server parsed infobox, a map from key to string or tuple
       *     null if server infobox is not valid
       */
      infobox?: Record<string, never>[]
      /**
       * Gender
       * @description parsed from wiki, maybe null
       */
      gender?: string
      /** @description parsed from wiki, maybe null, `1, 2, 3, 4` for `A, B, AB, O` */
      blood_type?: number & components["schemas"]["BloodType"]
      /**
       * Birth Year
       * @description parsed from wiki, maybe `null`
       */
      birth_year?: number
      /**
       * Birth Mon
       * @description parsed from wiki, maybe `null`
       */
      birth_mon?: number
      /**
       * Birth Day
       * @description parsed from wiki, maybe `null`
       */
      birth_day?: number
      stat: components["schemas"]["Stat"]
    }
// ...
    Episode: {
      /** ID */
      id: number
      /**
       * Type
       * @description `0` 本篇，`1` SP，`2` OP，`3` ED
       */
      type: number
      /** Name */
      name: string
      /** Name Cn */
      name_cn: string
      /**
       * Sort
       * @description 同类条目的排序和集数
       */
      sort: number
      /**
       * Ep
       * @description 条目内的集数, 从`1`开始。非本篇剧集的此字段无意义
       */
      ep?: number
      /** Airdate */
      airdate: string
      /** Comment */
      comment: number
      /**
       * Duration
       * @description 维基人填写的原始时长
       */
      duration: string
      /**
       * Desc
       * @description 简介
       */
      desc: string
      /**
       * Disc
       * @description 音乐曲目的碟片数
       */
      disc: number
      /** @description 服务器解析的时长，无法解析时为 `0` */
      duration_seconds?: number
    }
// ...
    WikiV0: {
      /** Key */
      key: string
      /** Value */
      value:
        | string
        | (
            | {
                /** K */
                k: string
                /** V */
                v: string
              }
            | {
                /** V */
                v: string
              }
          )[]
    }[]
// ...
  }
// ...
}